src/scripts/python-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai==1.3.0
google-generativeai==0.3.0
langchain==0.0.350
PyPDF2==3.0.1
python-pptx==0.6.21
nbformat==5.9.2
faiss-cpu==1.7.4
pydantic==2.5.0
